Refactor different run configurations into modules
The project started as a simple solution to run commands, but it quickly
grew into something big that needs to be flexible and support different configurations.

Every supported _task_ for a _hook type_ (`MFA`, `File`, `Command`,
`Mix task`...) is contained in its own module so this library will be
easy to extend and, more important, easy for outsiders to understand it
and collaborate.

Last, but not least, splitting the _tasks_ into module will improve the
documentation. Every particular case it will be explained in the
module's documentation page now.