Fix cmds chained with && are not treated separated
The problem is caused by the fact that the code uses a simple
`String.split(original_command, " ")` and treats the first token as
the executable (base_command) and the rest as arguments.

To allow using shell operators like `&&` or `||`, we can run
everything in a shell. So instead of splitting the command, we hand off
the entire string to the shell.